requirements.txt
aiohttp[speedups]==3.8.4
aiosqlite==0.19.0
discord.py[speed]==2.2.3
jishaku==2.5.1
quart==0.18.4
tortoise-orm==0.19.3
psutil==5.9.5
pillow==9.5.0